requirements.txt
elasticsearch==8.11.1
Flask==3.0.0
python-dotenv==1.0.0
